atoi actually returns a double

+++ IArguments.cc
@@ -30,7 +30,7 @@ int main (int argc, char* argv[])
   int const MyInt = atoi(argv[1]);
 
   // Get the double argument
-  float const MyFloat = atof(argv[2]);
+  double const MyFloat = atof(argv[2]);
 
   std::cout << "argv[1]: " << MyInt << std::endl;
   std::cout << "argv[2]: " << MyFloat << std::endl;
